refactor: LottoNumbersFactory 분리
